UI: Implement the working of components on the right hand screen
In the right hand screen, we have components: lock/unlock icon, current time,
outdoor temperature and username displayed

create a new folder controller in the same path as ui folder which holds
all the C++ classes which will be helpful to lookout for in the future.

Add new system C++ class which inherits from base class QObject and add all
the required QProperty(lock/unlock, current time, outdoor temperature and
username) to it.

Integrate the implemented C++ class to provide information to RightScreen.qml
file for the display of the components.